Refactor both endpoints into one file
